[RENAME] 이벤트 발행을 위한 VO 클래스 패키지 이동
