+++ src/main/java/io/oeid/mogakgo/common/event/domain/vo/AchievementCompletionEvent.java
@@ -1,16 +1,16 @@
-package io.oeid.mogakgo.common.event;
+package io.oeid.mogakgo.common.event.domain.vo;
 
 import io.oeid.mogakgo.domain.achievement.domain.entity.enums.ActivityType;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class AchievementNotificationEvent extends Event {
+public class AchievementCompletionEvent extends Event {
 
     private final Object target;
 
     @Builder
-    private AchievementNotificationEvent(Long userId, ActivityType activityType, Object target) {
+    private AchievementCompletionEvent(Long userId, ActivityType activityType, Object target) {
         super(userId, activityType);
         this.target = target;
     }
