+++ src/main/java/io/oeid/mogakgo/common/event/domain/vo/UserActivityEvent.java
@@ -1,14 +1,14 @@
-package io.oeid.mogakgo.common.event;
+package io.oeid.mogakgo.common.event.domain.vo;
 
 import io.oeid.mogakgo.domain.achievement.domain.entity.enums.ActivityType;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class UserActivityEvents extends Event {
+public class UserActivityEvent extends Event {
 
     @Builder
-    private UserActivityEvents(Long userId, ActivityType activityType) {
+    private UserActivityEvent(Long userId, ActivityType activityType) {
         super(userId, activityType);
     }
